refactor(utils): refactored delay to external function
